+++ drivers/starburst/src/metabase/driver/implementation/connectivity.clj
@@ -15,6 +15,8 @@
   "Connectivity implementation for Starburst driver."
   (:require [clojure.set :as set]
             [clojure.string :as str]
+            [clojure.tools.logging :as log]
+            [metabase.api.common :as api]
             [metabase.db.spec :as mdb.spec]
             [metabase.driver.sql-jdbc.common :as sql-jdbc.common]
             [metabase.driver.sql-jdbc.connection :as sql-jdbc.conn]
@@ -100,6 +102,17 @@
     (str v)
     v))
 
+; The role defined in the database should be ignored only if:
+; - The impersonation flag is checked AND
+; - The current user is NOT the database user
+(defn remove-role? [details-map]
+  (and
+    (:impersonation details-map)
+    (not= (get (deref api/*current-user*) :email) (:user details-map))))
+
+(defn assoc-if [coll key value condition]
+  (if condition (assoc coll key value) coll))
+
 (defmethod sql-jdbc.conn/connection-details->spec :starburst
   [_ details-map]
   (let [props (-> details-map
@@ -111,7 +124,10 @@
                   (update :kerberos str->bool)
                   (update :kerberos-delegation bool->str)
                   (assoc :SSL (:ssl details-map))
-                  (assoc :source "Starburst Metabase 2.0.0")
+                  (assoc :source "Starburst Metabase 6.1.0")
+                  (assoc-if :clientInfo "impersonate:true" (:impersonation details-map))
+                  (assoc-if :explicitPrepare "false" (:prepared-optimized details-map))
+                  (dissoc (if (remove-role? details-map) :roles :test))
 
                 ;; remove any Metabase specific properties that are not recognized by the Trino JDBC driver, which is
                 ;; very picky about properties (throwing an error if any are unrecognized)
@@ -125,6 +141,6 @@
                                  :source :applicationNamePrefix ::accessToken :SSL :SSLVerification :SSLKeyStorePath
                                  :SSLKeyStorePassword :SSLKeyStoreType :SSLTrustStorePath :SSLTrustStorePassword :SSLTrustStoreType :SSLUseSystemTrustStore
                                  :extraCredentials :roles :sessionProperties :externalAuthentication :externalAuthenticationTokenCache :disableCompression 
-                                 :assumeLiteralNamesInMetadataCallsForNonConformingClients]
+                                 :explicitPrepare :assumeLiteralNamesInMetadataCallsForNonConformingClients]
                                 (keys kerb-props->url-param-names))))]
     (jdbc-spec props)))