+++ drivers/starburst/src/metabase/driver/implementation/execute.clj
@@ -17,9 +17,14 @@
             [clojure.tools.logging :as log]
             [java-time :as t]
             [metabase.driver.implementation.sync :refer [starburst-type->base-type]]
+            [metabase.driver.implementation.messages :as msg]
+            [metabase.driver.sql-jdbc.connection :as sql-jdbc.conn]
             [metabase.driver.sql-jdbc.execute :as sql-jdbc.execute]
             [metabase.driver.sql.parameters.substitution :as sql.params.substitution]
             [metabase.query-processor.timezone :as qp.timezone]
+            [metabase.api.common :as api]
+            [metabase.query-processor.store :as qp.store]
+            [metabase.lib.metadata :as lib.metadata]
             [metabase.util.date-2 :as u.date]
             [metabase.util.i18n :refer [trs]])
   (:import com.mchange.v2.c3p0.C3P0ProxyConnection
@@ -41,27 +46,38 @@
   (-> (.. rs getStatement getConnection)
       pooled-conn->trino-conn))
 
-(defmethod sql-jdbc.execute/connection-with-timezone :starburst
-  [driver database ^String timezone-id]
-  ;; Trino supports setting the session timezone via a `TrinoConnection` instance method. Under the covers,
-  ;; https://github.com/trinodb/trino/blob/master/client/trino-jdbc/src/main/java/io/trino/jdbc/TrinoConnection.java#L596
-  ;; this is equivalent to the `X-Trino-Time-Zone` header in the HTTP request (i.e. the `:starburst` driver)
-  (let [conn            (.getConnection (sql-jdbc.execute/datasource-with-diagnostic-info! driver database))
-        underlying-conn (pooled-conn->trino-conn conn)]
-    (try
-      (sql-jdbc.execute/set-best-transaction-level! driver conn)
-      (when-not (str/blank? timezone-id)
-        ;; set session time zone if defined
-        (.setTimeZoneId underlying-conn timezone-id))
+(defmethod sql-jdbc.execute/do-with-connection-with-options :starburst
+  [driver db-or-id-or-spec {:keys [session-timezone write?], :as options} f]
+  (sql-jdbc.execute/do-with-resolved-connection
+    driver
+    db-or-id-or-spec
+    options
+    (fn [^java.sql.Connection conn]
       (try
-        (.setReadOnly conn true)
+        (sql-jdbc.execute/set-best-transaction-level! driver conn)
+        (let [underlying-conn (pooled-conn->trino-conn conn)
+              session-timezone (get options :session-timezone)]
+          (when-not (str/blank? (get options :session-timezone))
+            ;; set session time zone if defined
+            (.setTimeZoneId underlying-conn (get options :session-timezone))))
+        (try
+          (.setReadOnly conn true)
+          (catch Throwable e
+            (log/warn e (trs "Error setting Trino connection to read-only"))))
+          ;; as with statement and prepared-statement, cannot set holdability on the connection level
+        conn
         (catch Throwable e
-          (log/warn e (trs "Error setting Trino connection to read-only"))))
-      ;; as with statement and prepared-statement, cannot set holdability on the connection level
-      conn
-      (catch Throwable e
-        (.close conn)
-        (throw e)))))
+          (.close conn)
+          (throw e)))
+        (f conn))))
+
+(defmethod sql-jdbc.conn/data-warehouse-connection-pool-properties :starburst
+  [driver database]
+  ;; In order to avoid Metabase from cancelling long queries, the
+  ;; unreturnedConnectionTimeout property needs to be updated
+  (merge
+   ((get-method sql-jdbc.conn/data-warehouse-connection-pool-properties :sql-jdbc) driver database)
+   {"unreturnedConnectionTimeout" 100000000000}))
 
 ;;; +----------------------------------------------------------------------------------------------------------------+
 ;;; |                                          Reading Columns from Result Set                                       |
@@ -144,10 +160,113 @@
 ;;; |                                          SQL Statment Operations                                               |
 ;;; +----------------------------------------------------------------------------------------------------------------+
 
+(defn impersonate-user
+  [conn]
+  (if
+    (clojure.string/includes? (.getProperty (.getClientInfo conn) "ClientInfo" "") "impersonate:true")
+    (let [email (get (deref api/*current-user*) :email)]
+      (.setSessionUser (.unwrap conn TrinoConnection) email))
+    nil))
+
+(defn remove-impersonation
+  [conn]
+  (if
+    (clojure.string/includes? (.getProperty (.getClientInfo conn) "ClientInfo" "") "impersonate:true")
+    (.clearSessionUser (.unwrap conn TrinoConnection))
+    nil))
+
+; Metabase tests require a specific error when an invalid number of parameters are passed
+(defn handle-execution-error
+  [e]
+  (let [message (.getMessage e)]
+    (cond
+      (clojure.string/includes? message "Expecting: 'USING'")
+      (throw (Exception. (str message msg/STARBURST_MAYBE_INCOMPATIBLE)))
+      (clojure.string/includes? message "Incorrect number of parameters")
+      (throw (Exception. msg/TOO_MANY_PARAMETERS))
+      :else (throw e))))
+
+; Optimized prepared statement where a proxy is generated and set-parameters! called on that proxy.
+; Metabase is sometimes calling getParametersMetaData() on the prepared statement in order to count
+; the number of parameters and verify they are correct with what is expected
+; This unfortunately defeats the purpose of using EXECUTE IMMEDIATE as it forces the JDBC driver
+; to call an explicit PREPARE. However this call is optional, so the solution is to create a proxy
+; which defines all methods used by Metabase *except* for metadata methods. When Metabase tries
+; to count the number of parameters:
+; - The proxy issues an exception as getParametersMetaData() is not defined
+; - Metabase catches the exception and does not perform the check
+; - An invalid query is sent to Trino, which fails with a "Incorrect number of parameters" message
+; - This message is caught by the driver and replaced with the exact same Metabase message
+; In the end, the exact same message is presented to the user when the number of arguments is
+; incorrect except we now execute the query to display the error message
+(defn proxy-optimized-prepared-statement
+  [driver conn stmt params]
+  (let [ps (proxy [java.sql.PreparedStatement] []
+          (executeQuery []
+            (try
+              (let [rs (.executeQuery stmt)]
+                (remove-impersonation conn)
+                rs)
+                (catch Throwable e (handle-execution-error e))))
+          (execute []
+            (try
+              (let [rs (.execute stmt)]
+                (remove-impersonation conn)
+                rs)
+                (catch Throwable e (handle-execution-error e))))
+          (setMaxRows [nb] (.setMaxRows stmt nb))
+          (setObject
+            ([index obj] (.setObject stmt index obj))
+            ([index obj type] (.setObject stmt index obj type)))
+          (setTime
+            ([index val] (.setTime stmt index val))
+            ([index val cal] (.setTime stmt index val cal)))
+          (setTimestamp
+            ([index val] (.setTimestamp stmt index val))
+            ([index val cal] (.setTimestamp stmt index val cal)))
+          (setDate
+            ([index val] (.setDate stmt index val))
+            ([index val cal] (.setDate stmt index val cal)))
+          (setArray [index val] (.setArray stmt index val))
+          (setBoolean [index val] (.setBoolean stmt index val))
+          (setByte [index val] (.setByte stmt index val))
+          (setBytes [index val] (.setBytes stmt index val))
+          (setInt [index val] (.setInt stmt index val))
+          (setShort [index val] (.setShort stmt index val))
+          (setLong [index val] (.setLong stmt index val))
+          (setFloat [index val] (.setFloat stmt index val))
+          (setDouble [index val] (.setDouble stmt index val))
+          (cancel [] (.cancel stmt))
+          (close [] (.close stmt)))]
+    (sql-jdbc.execute/set-parameters! driver ps params)
+    ps))
+
+; Default prepared statement where set-parameters! is called before generating the proxy
+(defn proxy-prepared-statement
+  [driver conn stmt params]
+  (sql-jdbc.execute/set-parameters! driver stmt params)
+  (proxy [java.sql.PreparedStatement] []
+    (executeQuery []
+      (try
+        (let [rs (.executeQuery stmt)]
+          (remove-impersonation conn)
+          rs)
+          (catch Throwable e (handle-execution-error e))))
+    (execute []
+      (try
+        (let [rs (.execute stmt)]
+          (remove-impersonation conn)
+          rs)
+          (catch Throwable e (handle-execution-error e))))
+    (setMaxRows [nb] (.setMaxRows stmt nb))
+    (cancel [] (.cancel stmt))
+    (close [] (.close stmt))))
+
 (defmethod sql-jdbc.execute/prepared-statement :starburst
   [driver ^Connection conn ^String sql params]
   ;; with Starburst driver, result set holdability must be HOLD_CURSORS_OVER_COMMIT
   ;; defining this method simply to omit setting the holdability
+  (impersonate-user conn)
   (let [stmt (.prepareStatement conn
                                 sql
                                 ResultSet/TYPE_FORWARD_ONLY
@@ -157,16 +276,19 @@
         (.setFetchDirection stmt ResultSet/FETCH_FORWARD)
         (catch Throwable e
           (log/debug e (trs "Error setting prepared statement fetch direction to FETCH_FORWARD"))))
-      (sql-jdbc.execute/set-parameters! driver stmt params)
-      stmt
+      (if
+        (.useExplicitPrepare (.unwrap conn TrinoConnection))
+        (proxy-prepared-statement driver conn stmt params)
+        (proxy-optimized-prepared-statement driver conn stmt params))
       (catch Throwable e
+        (remove-impersonation conn)
         (.close stmt)
         (throw e)))))
-
 
 (defmethod sql-jdbc.execute/statement :starburst
   [_ ^Connection conn]
   ;; and similarly for statement (do not set holdability)
+  (impersonate-user conn)
   (let [stmt (.createStatement conn
                                ResultSet/TYPE_FORWARD_ONLY
                                ResultSet/CONCUR_READ_ONLY)]
@@ -174,7 +296,17 @@
       (.setFetchDirection stmt ResultSet/FETCH_FORWARD)
       (catch Throwable e
         (log/debug e (trs "Error setting statement fetch direction to FETCH_FORWARD"))))
-    stmt))
+    (proxy [java.sql.Statement] []
+      (execute [sql]
+        (try
+          (let [rs (.execute stmt sql)]
+            rs)
+            (catch Throwable e (handle-execution-error e))
+            (finally (remove-impersonation conn))))
+      (getResultSet [] (.getResultSet stmt))
+      (setMaxRows [nb] (.setMaxRows stmt nb))
+      (cancel [] (.cancel stmt))
+      (close [] (.close stmt)))))
 
 ;;; +----------------------------------------------------------------------------------------------------------------+
 ;;; |                                          Prepared Statement Substitutions                                      |
@@ -189,7 +321,7 @@
   ;; (which was set via report time zone), it is necessary to use the `from_iso8601_timestamp` function on the string
   ;; representation of the `ZonedDateTime` instance, but converted to the report time zone
   ;_(date-time->substitution (.format (t/offset-date-time (t/local-date-time t) (t/zone-offset 0)) DateTimeFormatter/ISO_OFFSET_DATE_TIME))
-  (let [report-zone       (qp.timezone/report-timezone-id-if-supported :starburst)
+  (let [report-zone       (qp.timezone/report-timezone-id-if-supported :starburst (lib.metadata/database (qp.store/metadata-provider)))
         ^ZonedDateTime ts (if (str/blank? report-zone) t (t/with-zone-same-instant t (t/zone-id report-zone)))]
     ;; the `from_iso8601_timestamp` only accepts timestamps with an offset (not a zone ID), so only format with offset
     (date-time->substitution (.format ts DateTimeFormatter/ISO_OFFSET_DATE_TIME))))