+++ drivers/starburst/src/metabase/driver/implementation/sync.clj
@@ -20,11 +20,13 @@
             [java-time :as t]
             [metabase.driver :as driver]
             [metabase.driver.sql-jdbc.connection :as sql-jdbc.conn]
-            [metabase.driver.sql-jdbc.sync :as sql-jdbc.sync]
             [metabase.driver.sql-jdbc.sync.describe-database :as sql-jdbc.describe-database]
             [metabase.driver.sql-jdbc.execute :as sql-jdbc.execute]
+            [metabase.driver.sql-jdbc.sync.interface :as sql-jdbc.sync.interface]
+            [metabase.driver.sql-jdbc.sync :as sql-jdbc.sync]
             [metabase.driver.sql.util :as sql.u]
-            [metabase.util.i18n :refer [trs]]))
+            [metabase.util.i18n :refer [trs]])
+  (:import (java.sql Connection)))
 
 (def starburst-type->base-type
   "Function that returns a `base-type` for the given `straburst-type` (can be a keyword or string)."
@@ -91,15 +93,16 @@
     (log/debugf "database-type->base-type %s -> %s" field-type base-type)
     base-type))
 
-(defn- have-select-privilege?
-  "Checks whether the connected user has permission to select from the given `table-name`, in the given `schema`."
-  [driver conn schema table-name]
+(defmethod sql-jdbc.sync.interface/have-select-privilege? :starburst
+  [driver ^Connection conn table-schema table-name]
   (try
-    (let [sql (sql-jdbc.describe-database/simple-select-probe-query driver schema table-name)]
-        ;; if the query completes without throwing an Exception, we can SELECT from this table
-      (jdbc/reducible-query {:connection conn} sql)
-      true)
-    (catch Throwable _
+    (let [sql (str "SHOW TABLES FROM \"" table-schema "\" LIKE '" table-name "'")]
+      ;; if the query completes without throwing an Exception, we can SELECT from this table
+      (with-open [stmt (.prepareStatement conn sql)
+                  rs (.executeQuery stmt)]
+          (.next rs)))
+    (catch Throwable e
+      (log/fatal "ERROR WITH QUERY " e)
       false)))
 
 (defn- describe-schema
@@ -111,7 +114,7 @@
       (into
        #{}
        (comp (filter (fn [{table-name :table}]
-                                (have-select-privilege? driver conn schema table-name)))
+                                (sql-jdbc.sync.interface/have-select-privilege? driver conn schema table-name)))
                       (map (fn [{table-name :table}]
                              {:name        table-name
                               :schema      schema})))
@@ -138,46 +141,44 @@
 
 (defmethod driver/describe-database :starburst
   [driver {{:keys [catalog schema] :as details} :details :as database}]
-  (with-open [conn (-> (sql-jdbc.conn/db->pooled-connection-spec database)
-                       jdbc/get-connection)]
-    (let [schemas (if schema #{(describe-schema driver conn catalog schema)}
+  (sql-jdbc.execute/do-with-connection-with-options
+    driver
+    database
+    nil
+    (fn [^Connection conn]
+      (let [schemas (if schema #{(describe-schema driver conn catalog schema)}
                       (all-schemas driver conn catalog))]
-      {:tables (reduce set/union schemas)})))
+        {:tables (reduce set/union schemas)}))))
 
 (defmethod driver/describe-table :starburst
   [driver {{:keys [catalog] :as details} :details :as database} {schema :schema, table-name :name}]
-  (with-open [conn (-> (sql-jdbc.conn/db->pooled-connection-spec database)
-                       jdbc/get-connection)
-              stmt (.createStatement conn)]
-    (let [sql (describe-table-sql driver catalog schema table-name)
-          rs (sql-jdbc.execute/execute-statement! driver stmt sql)
-          show-create-table-sql (show-create-table-sql driver catalog schema table-name)
-          partitioned_by (extract-partitioned-by (jdbc/result-set-seq (sql-jdbc.execute/execute-statement! driver stmt show-create-table-sql)))]
-      {:schema schema
-       :name   table-name
-       :fields (into
-                #{}
-                (map-indexed (fn [idx {:keys [column type extra] :as col}]
-                                 (if (some #(= (str "'" column "'") %) partitioned_by)
-                                   {:name column
-                                    :database-type type
-                                    :base-type         (starburst-type->base-type type)
-                                    :database-position idx
-                                    :field-comment "partition key"
-                                    }
-                                   {:name column
-                                    :database-type type
-                                    :base-type         (starburst-type->base-type type)
-                                    :database-position idx
-                                    })
-                               ))
-                (jdbc/reducible-query {:connection conn} sql))})))
+  (sql-jdbc.execute/do-with-connection-with-options
+    driver
+    database
+    nil
+    (fn [^Connection conn]
+      (with-open [stmt (.createStatement conn)]
+        (let [sql (describe-table-sql driver catalog schema table-name)
+              rs (sql-jdbc.execute/execute-statement! driver stmt sql)]
+          {:schema schema
+           :name   table-name
+           :fields (into
+            #{}
+            (map-indexed (fn [idx {:keys [column type] :as col}]
+                            {:name column
+                            :database-type type
+                            :base-type         (starburst-type->base-type type)
+                            :database-position idx}))
+            (jdbc/reducible-result-set rs {}))})))))
 
 (defmethod driver/db-default-timezone :starburst
   [driver {{:keys [catalog] :as details} :details :as database}]
-  (with-open [conn (-> (sql-jdbc.conn/db->pooled-connection-spec database)
-                       jdbc/get-connection)
-              stmt (.createStatement conn)]
-    (let [rs (sql-jdbc.execute/execute-statement! driver stmt "SELECT current_timezone() as \"time-zone\"")
-          [{:keys [time-zone]}] (jdbc/result-set-seq rs)]
-      time-zone)))
+  (sql-jdbc.execute/do-with-connection-with-options
+    driver
+    database
+    nil
+    (fn [^Connection conn]
+      (with-open [stmt (.createStatement conn)]
+        (let [rs (sql-jdbc.execute/execute-statement! driver stmt "SELECT current_timezone() as \"time-zone\"")
+              [{:keys [time-zone]}] (jdbc/result-set-seq rs)]
+          time-zone)))))