+++ drivers/starburst/src/metabase/driver/starburst.clj
@@ -14,13 +14,53 @@
 (ns metabase.driver.starburst
   "Starburst driver."
   (:require [metabase.driver :as driver]
+            [clojure.java.jdbc :as jdbc]
+            [clojure.tools.logging :as log]
+            [metabase.util.ssh :as ssh]
+            [metabase.driver.sql-jdbc.connection :as connection]
+            [metabase.query-processor.util :as qp.util]
+            [metabase.public-settings :as public-settings]
+            [metabase.driver.implementation.messages :as msg]
             [metabase.driver.sql-jdbc.execute.legacy-impl :as sql-jdbc.legacy]))
-
-(set! *warn-on-reflection* true)
-
 (driver/register! :starburst, :parent #{::sql-jdbc.legacy/use-legacy-classes-for-read-and-set})
  
-(prefer-method driver/supports? [:starburst :set-timezone] [:sql-jdbc :set-timezone])
+(prefer-method driver/database-supports? [:starburst :set-timezone] [:sql-jdbc :set-timezone])
+
+(defn format-field
+  [name value]
+  (if (nil? value)
+    ""
+    (str " " name ": " value)))
+
+(defmethod qp.util/query->remark :starburst
+  [_ {{:keys [card-id dashboard-id]} :info, :as query}]
+  (str
+    (qp.util/default-query->remark query)
+    (format-field "accountID" (public-settings/site-uuid))
+    (format-field "dashboardID" dashboard-id)
+    (format-field "cardID" card-id)))
+
+(defn handle-execution-error
+  [e details]
+  (let [message (.getMessage e)
+        execute-immediate (get details :prepared-optimized false)]
+    (cond
+      (and (clojure.string/includes? message "Expecting: 'USING'") execute-immediate)
+      (throw (Exception. msg/STARBURST_INCOMPATIBLE_WITH_OPTIMIZED_PREPARED))
+      :else (throw e))))
+
+(defmethod driver/can-connect? :starburst
+  [driver details]
+  (try
+    (connection/with-connection-spec-for-testing-connection [jdbc-spec [driver details]]
+      (connection/can-connect-with-spec? jdbc-spec))
+    (catch Throwable e (handle-execution-error e details))))
+
+;;; The Starburst JDBC driver DOES NOT support the `.getImportedKeys` method so just return `nil` here so the
+;;; implementation doesn't try to use it.
+(defmethod driver/describe-table-fks :starburst
+  [_driver _database _table]
+  nil)
 
 ;;; +----------------------------------------------------------------------------------------------------------------+
 ;;; |                                                  Load implemetation files                                      |